+++ nodeconductor/server/admin/dashboard.py
@@ -1,9 +1,9 @@
 from django.utils.translation import ugettext_lazy as _
 from fluent_dashboard.dashboard import modules, FluentIndexDashboard, FluentAppIndexDashboard
-from fluent_dashboard.modules import AppIconList
+from fluent_dashboard.modules import AppIconList, PersonalModule
 
 from nodeconductor.core import NodeConductorExtension
-
+from nodeconductor import __version__
 
 class CustomIndexDashboard(FluentIndexDashboard):
     """
@@ -11,19 +11,47 @@ class CustomIndexDashboard(FluentIndexDashboard):
     """
     title = 'NodeConductor administration'
 
+    def _get_installed_plugin_info(self):
+        result = [
+            {
+                'title': _('NodeConductor %s' % __version__),
+                'url': 'http://nodeconductor.readthedocs.org/en/stable/',
+                'external': True,
+            },
+        ]
+
+        documentation = {
+            'nodeconductor_organization': 'http://nodeconductor-organization.readthedocs.org/en/latest/',
+            'nodeconductor_saltstack': 'http://nodeconductor-saltstack.readthedocs.org/',
+            'nodeconductor_sugarcrm': 'http://nodeconductor-sugarcrm.readthedocs.org/',
+
+        }
+        plugins = set()
+        for ext in NodeConductorExtension.get_extensions():
+            base_name = ext.__module__.split('.')[0]
+            module = __import__(base_name)
+            name = 'NodeConductor %s' % ' '.join(base_name.split('_')[1:]).title()
+            plugins.add((name, getattr(module, '__version__', 'N/A'), module.__doc__))
+
+        for plugin, version, description in sorted(plugins):
+            result.append(
+                {
+                    'title': '%s %s' % (plugin, version),
+                    'url': documentation.get(plugin,
+                                             'http://nodeconductor.readthedocs.org/en/latest/#nodeconductor-plugins'),
+                    'description': description,
+                    'external': True
+                }
+            )
+
+        return result
+
     def __init__(self, **kwargs):
         FluentIndexDashboard.__init__(self, **kwargs)
         self.children.append(modules.Group(
             title="IaaS",
             display="tabs",
             children=[
-                modules.ModelList(
-                    title='Network',
-                    models=('nodeconductor.iaas.models.FloatingIP',
-                            'nodeconductor.iaas.models.IpMapping',
-                            'nodeconductor.iaas.models.SecurityGroup',
-                            )
-                ),
                 modules.ModelList(
                     title='Virtual Machine',
                     models=('nodeconductor.iaas.models.Instance',
@@ -38,7 +66,14 @@ class CustomIndexDashboard(FluentIndexDashboard):
                             'nodeconductor.iaas.models.CloudProjectMembership',
                             'nodeconductor.iaas.models.OpenStackSettings',
                             )
-                )
+                ),
+                modules.ModelList(
+                    title='Network',
+                    models=('nodeconductor.iaas.models.FloatingIP',
+                            'nodeconductor.iaas.models.IpMapping',
+                            'nodeconductor.iaas.models.SecurityGroup',
+                            )
+                ),
             ]
         ))
 
@@ -49,20 +84,19 @@ class CustomIndexDashboard(FluentIndexDashboard):
             billing_models.append('nodeconductor_paypal.models.Payment')
 
         self.children.append(AppIconList(_('Billing'), models=billing_models))
-        self.children.append(AppIconList(_('Structure'), models=('nodeconductor.structure.*',)))
-        self.children.append(AppIconList(_('Backup'), models=('nodeconductor.backup.*',)))
+        self.children.append(AppIconList(_('Structure'), models=(
+            'nodeconductor.structure.*',
+            'nodeconductor.core.models.User',
+        )))
 
     def init_with_context(self, context):
         self.children.append(modules.LinkList(
-            _('Quick links'),
-            layout='inline',
-            draggable=False,
-            deletable=False,
-            collapsible=False,
-            children=[
-                [_('API'), '/api/'],
-                [_('Documentation'), 'http://nodeconductor.readthedocs.org/en/stable/'],
-            ]
+            _('Installed components'),
+            layout='stacked',
+            draggable=True,
+            deletable=True,
+            collapsible=True,
+            children=self._get_installed_plugin_info()
         ))